Move `FormatPage` tests into a dedicated test module
