+++ test-prototypes/Main.hs
@@ -2,13 +2,13 @@ module Main (main) where
 
 import           Test.Tasty
 
-import qualified FormatPage
+import qualified Test.FormatPage
 import qualified Test.ScheduledMerges
 import qualified Test.ScheduledMergesQLS
 
 main :: IO ()
 main = defaultMain $ testGroup "prototypes" [
-      FormatPage.tests
+      Test.FormatPage.tests
     , Test.ScheduledMerges.tests
     , Test.ScheduledMergesQLS.tests
     ]
